+++ common/src/main/java/dev/aika/i_have_slept/IHaveSlept.java
@@ -1,5 +1,7 @@
 package dev.aika.i_have_slept;
 
+import dev.aika.i_have_slept.mixin.GameRulesAccessor;
+import dev.aika.i_have_slept.mixin.GameRulesBooleanValueAccessor;
 import net.minecraft.world.level.GameRules;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,5 +14,10 @@ public final class IHaveSlept {
     public static GameRules.Key<GameRules.BooleanValue> BETTER_SERVER_SLEEP;
 
     public static void init() {
+        BETTER_SERVER_SLEEP = GameRulesAccessor.callRegister(
+                "doBetterServerSleep",
+                GameRules.Category.PLAYER,
+                GameRulesBooleanValueAccessor.invokeCreate(true)
+        );
     }
 }