+++ CHANGELOG.md
@@ -1,1 +1,12 @@
 ### Unreleased
+
+**NeoForge**
+
+- add gamerule translations
+- move gamerule doBetterServerSleep to player category
+
+**Fabric**
+
+- add gamerule translations
+- remove fabric-api dependency
+- move gamerule doBetterServerSleep to player category